chore(build): adjusted cleanup scripts
To make it easier to push to `gh-pages`, the `dist` folder (which is ignored by
the main branch) is being used as the root of the `gh-pages` branch. The old
cleaning scripts deleted it altogether, which wiped out the `.git` dir and made
things difficult. So now they just delete contents, not the whole dir.